+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next release...
+* Gentoo with musl use an additional repo https://github.com/gentoo/musl.git.
 * Gentoo use the kernel `sys-kernel/gentoo-kernel-bin` to install more quickly.
 * No more need to enter password twice with Grub and encrypted system.
 * Can restart the whole installation from scratch with the option `--restart`